refactor step1
