+++ packages/light-apps/package.json
@@ -20,10 +20,9 @@
     "@polkadot/transfer-app": "0.1.0",
     "@polkadot/ui-assets": "^0.23.1",
     "@polkadot/ui-components": "0.1.0",
-    "@polkadot/ui-keyring": "^0.24.2",
-    "@polkadot/ui-settings": "^0.22.17",
-    "@polkadot/util": "^0.33.20",
-    "@polkadot/util-crypto": "^0.33.20",
+    "@polkadot/ui-keyring": "^0.25.7",
+    "@polkadot/util": "^0.33.35",
+    "@polkadot/util-crypto": "^0.33.35",
     "@types/file-saver": "^2.0.0",
     "@types/react-dom": "^16.0.11",
     "@types/react-router-dom": "^4.3.1",
